ui/control/RichTextWindow: Advance focus after Enter on checkboxes
Move FocusItem into the header and add AdvanceFocusToNextFrom() so
KEY_RETURN on a focused checkbox toggles it then applies the same
rules as KEY_DOWN (next item, max_jump, end-of-list / parent inject).

Improves keyboard flow in checklist-style rich text (e.g. Checklist dialog).

Fixes #2405
