+++ NEWS.txt
@@ -28,6 +28,7 @@ Version 7.45 - not yet released
   - tabbed dialogs: show list-style focus on the tab strip when the device has
     cursor keys and the strip has keyboard focus (or always on devices without
     cursor keys, where the indicator is informational only)
+  - checklist rich text: Enter on a checkbox advances focus like Down #2405
 * android
   - use physical display metrics for startup DPI (getRealMetrics) #1784
   - UI sounds via preloaded SoundPool (MediaPlayer fallback for early play or
